+++ src/models/TicketStatus.js
@@ -4,9 +4,24 @@ class TicketStatus extends Model {
   constructor (data) {
     super(data, [
       { key: 'id' },
-      { key: 'name' },
+      { key: 'name' }, // unanswered - answered - pending - closed
       { key: 'title' }
     ])
+  }
+
+  getStatusColor () {
+    switch (this.name) {
+      case 'answered':
+        return 'positive'
+      case 'unanswered':
+        return 'negative'
+      case 'pending':
+        return 'warning'
+      case 'closed':
+        return 'info'
+      default:
+        return 'info'
+    }
   }
 }